rename software package
