+++ eventmesh-console/src/main/java/org/apache/eventmesh/admin/config/AdminProperties.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.console.config;
+package org.apache.eventmesh.admin.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
