+++ eventmesh-console/src/main/java/org/apache/eventmesh/admin/config/Constants.java
@@ -15,8 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.console.model;
+package org.apache.eventmesh.admin.config;
 
-public class ConnectionInfo {
+public class Constants {
 
+    // config
+    public static final String ADMIN_PROPS_PREFIX = "eventmesh";
+    public static final String META_TYPE_NACOS = "nacos";
+    public static final String META_TYPE_ETCD = "etcd";
+
+    // Open-API
+    public static final String HTTP_PREFIX = "http://";
+    public static final String HTTPS_PREFIX = "https://";
+
+    // Nacos
+    public static final String NACOS_LOGIN_API = "/nacos/v1/auth/login";
+    public static final String NACOS_CONFIGS_API = "/nacos/v1/cs/configs";
 }
